requirements.txt
groq==0.11.0
langchain==0.3.0
langchain-groq==0.2.0
langserve==0.3.0
pydantic==2.9.2
python-decouple==3.8
uvicorn==0.30.6
